fix: properly scaling eval and perspective
Eval phase is now scaled from [0.0, 1.0] while parsing so we don't need the max game phase while computing in the tuner. Also made sure that tuning evaluation is from white's perspective to be consistent with how to evaluate positions and the tuner scores. Added comments to make this more clear as well.